WIP: UserCredentials: explain how to get a User token after logging in
This class tries to use either the passed-in token, or the auth code to
log in and get increased quota or permissions (Scopes).

I think the code should only be used once, hence it is not as good for
our experiments at the moment, but the token can be reused for some time.